raftstore: extract some coprocessor traits
