+++ src/raftstore/coprocessor/dispatcher.rs
@@ -5,6 +5,7 @@ use kvproto::metapb::Region;
 use kvproto::pdpb::CheckPolicy;
 use kvproto::raft_cmdpb::{RaftCmdRequest, RaftCmdResponse};
 
+use raft::StateRole;
 use raftstore2::store::transport::CasualRouter;
 
 use super::*;
